refactor: apply loading toast message (#3800)
* refactor: update loading toast message


* refactor: apply loading toast message after modal closed


* chore: delete setTimeout


---------
